trade_bot_binance_v0.001/requirements.txt
fastapi
uvicorn
httpx
sqlalchemy
psycopg2-binary
python-dotenv
python-binance
aiogram
pandas
numpy
scikit-learn
joblib
openai
requests
numpy
matplotlib
seaborn
statsmodels
